+++ src/fancy/IMagickImage.h
@@ -25,9 +25,9 @@ public:
 								size_t square_side, 
 								const std::string& suffix);
 									
-	Magick::Color process_element(unsigned int x0, unsigned int y0, unsigned int x1, unsigned int y1);
+	Magick::Color process_element(size_t x0, size_t y0, size_t x1, size_t y1);
 	
-	// static std::vector<Magick::Color> getColors(unsigned int num_colors, unsigned int max_color);
+	// static std::vector<Magick::Color> getColors(size_t num_colors, size_t max_color);
 	
 private:
 	IMagickImage(const std::string& filename);
