+++ src/types/color/Color.cpp
@@ -15,7 +15,7 @@ Color::Color() :
 	_r(0), _g(0), _b(0)
 {}
 
-Color::Color(unsigned int r, unsigned int g, unsigned int b) :
+Color::Color(size_t r, size_t g, size_t b) :
 	_r(r), _g(g), _b(b)
 {}
 
@@ -31,9 +31,9 @@ Color::Color(const std::string& str, const std::string& delimiter)
 	
 	try
 	{
-		_r = boost::lexical_cast<unsigned int>(rgb[0]);
-		_g = boost::lexical_cast<unsigned int>(rgb[1]);
-		_b = boost::lexical_cast<unsigned int>(rgb[2]);
+		_r = boost::lexical_cast<size_t>(rgb[0]);
+		_g = boost::lexical_cast<size_t>(rgb[1]);
+		_b = boost::lexical_cast<size_t>(rgb[2]);
 	}
 	catch(boost::bad_lexical_cast& exc)
 	{
@@ -62,8 +62,8 @@ std::string Color::str() const
 
 bool operator< (const Color& c1, const Color& c2)
 {
-	unsigned int sum1(c1._r + c1._g + c1._b);
-	unsigned int sum2(c2._r + c2._g + c2._b);
+	size_t sum1(c1._r + c1._g + c1._b);
+	size_t sum2(c2._r + c2._g + c2._b);
 	
 	return sum1 < sum2;
 }
